+++ eyes.selenium.java/src/test/java/com/applitools/eyes/selenium/TestVisualLocators.java
@@ -1,10 +1,8 @@
 package com.applitools.eyes.selenium;
 
-import com.applitools.eyes.EyesRunner;
-import com.applitools.eyes.FixedCutProvider;
-import com.applitools.eyes.Region;
-import com.applitools.eyes.StdoutLogHandler;
+import com.applitools.eyes.*;
 import com.applitools.eyes.locators.VisualLocator;
+import com.applitools.eyes.selenium.fluent.Target;
 import com.applitools.eyes.utils.ReportingTestSuite;
 import com.applitools.eyes.utils.SeleniumUtils;
 import com.applitools.eyes.visualgrid.services.VisualGridRunner;
@@ -35,13 +33,18 @@ public class TestVisualLocators extends ReportingTestSuite {
         String suffix = useVisualGrid ? "_VG" : "";
         Eyes eyes = new Eyes(runner);
         eyes.setLogHandler(new StdoutLogHandler());
+        eyes.setSaveNewTests(false);
         RemoteWebDriver driver = SeleniumUtils.createChromeDriver();
         driver.get("https://applitools.github.io/demo/TestPages/FramesTestPage/");
         try {
-            eyes.open(driver, "Applitools Eyes SDK", "testVisualLocators" + suffix);
+            eyes.open(driver, "Applitools Eyes SDK", "testVisualLocators" + suffix, new RectangleSize(800, 600));
+            eyes.check(Target.window().fully(false));
             Map<String, List<Region>> result = eyes.locate(VisualLocator.name("applitools_title"));
             eyes.setImageCut(new FixedCutProvider(19, 0, 3, 0));
             Map<String, List<Region>> resultCut = eyes.locate(VisualLocator.name("applitools_title"));
+            eyes.setImageCut(null);
+            eyes.setScaleRatio(0.5);
+            Map<String, List<Region>> resultScale = eyes.locate(VisualLocator.name("applitools_title_scaled_down"));
             eyes.closeAsync();
 
             Assert.assertEquals(result.size(), 1);
@@ -55,6 +58,12 @@ public class TestVisualLocators extends ReportingTestSuite {
             Assert.assertEquals(regionList.size(), 1);
             region = regionList.get(0);
             Assert.assertEquals(region, new Region(0, 0, 158, 38));
+
+            Assert.assertEquals(resultScale.size(), 1);
+            regionList = resultScale.get("applitools_title_scaled_down");
+            Assert.assertEquals(regionList.size(), 1);
+            region = regionList.get(0);
+            Assert.assertEquals(region, new Region(2, 8, 77, 22));
         } finally {
             driver.quit();
             eyes.abortAsync();